# backend/src/docs/API/api.yaml
openapi: 3.0.1
info:
  title: AI-Powered Knowledge Management API
  description: REST API for managing AI-assisted knowledge entries and categories.
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
parameters:
  refreshTokenCookie:
    name: refreshToken
    in: cookie
    description: Refresh token for authentication
    required: false
    schema:
      type: string
paths:
  /api/notes:
    get:
      security:
        - bearerAuth: []
      summary: Get all knowledge entries
      description: Returns a list of knowledge entries. You can filter by archived status and category.
      parameters:
        - in: query
          name: archived
          schema:
            type: boolean
          description: Filter entries by archived status (true/false)
        - in: query
          name: category
          schema:
            type: integer
          description: Filter entries by category ID. Use -1 to filter by no category or leave empty to get all entries.
      responses:
        200:
          description: List of knowledge entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      security:
        - bearerAuth: []
      summary: Create a new knowledge entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        201:
          description: Knowledge entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: Invalid data for creating a knowledge entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notes/{id}:
    delete:
      security:
        - bearerAuth: []
      summary: Delete a knowledge entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Knowledge entry deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: ID is required to delete a knowledge entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Knowledge entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      security:
        - bearerAuth: []
      summary: Update a knowledge entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdateInput'
      responses:
        200:
          description: Knowledge entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: Invalid data for updating a knowledge entry or ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Knowledge entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notes/search:
    post:
      security:
        - bearerAuth: []
      summary: Run semantic search across knowledge entries
      description: Returns up to 10 knowledge entries ranked by cosine similarity to the provided query.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Natural language prompt to search for semantically related entries.
              required:
                - query
      responses:
        200:
          description: Ranked knowledge entries matching the semantic query
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
        400:
          description: Invalid search request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notes/{id}/summarize:
    post:
      security:
        - bearerAuth: []
      summary: Summarize a knowledge entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Concise summary generated by Google Generative AI
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: object
                    properties:
                      summary:
                        type: string
        400:
          description: Missing required data to summarize
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Knowledge entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories:
    get:
      security:
        - bearerAuth: []
      summary: Get all categories
      responses:
        200:
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      security:
        - bearerAuth: []
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        201:
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Invalid data for creating category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories/{id}:
    delete:
      security:
        - bearerAuth: []
      summary: Delete a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: ID is required to delete category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      security:
        - bearerAuth: []
      summary: Update a category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        200:
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Invalid data for updating category or ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/signup:
    post:
      summary: Create a new user
      description: Registers a new user in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid data or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/sessions/login:
    post:
      summary: Log in
      description: Logs in and returns the accessToken and refreshToken.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: integer
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/sessions/refresh:
    post:
      security:
        - bearerAuth: []
      summary: Refresh token
      description: Returns a new accessToken using the refreshToken.
      parameters:
        - $ref: '#/parameters/refreshTokenCookie'
      responses:
        200:
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: integer
        401:
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/logout:
    post:
      security:
        - bearerAuth: []
      summary: Log out
      description: Deletes the refreshToken and logs out the user.
      parameters:
        - $ref: '#/parameters/refreshTokenCookie'
      responses:
        200:
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: integer
          nullable: true
        archived:
          type: boolean
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
          nullable: true
        Category:
          $ref: '#/components/schemas/Category'
        similarityScore:
          type: number
          format: float
          nullable: true
          description: Cosine similarity score returned by semantic search queries.
    NoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: integer
          nullable: true
      required:
        - title
        - content
    NoteUpdateInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: integer
          nullable: true
        archived:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
          nullable: true
        Notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
    CategoryInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Error:
      type: object
      properties:
        status:
          type: string
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string