# backend/src/models/note.model.js
import { DataTypes } from 'sequelize';
import { sequelize } from '../config/database.js';

export const Note = sequelize.define('Note', {
    id: {
        type: DataTypes.INTEGER,
        primaryKey: true,
        autoIncrement: true
    },
    title: {
        type: DataTypes.STRING,
        allowNull: false
    },
    content: {
        type: DataTypes.TEXT,
        allowNull: false
    },
    category: {
        type: DataTypes.INTEGER,
        allowNull: true,
        references: {
            model: 'Categories',
            key: 'id'
        }
    },
    embedding: {
        type: DataTypes.JSON,
        allowNull: true
    },
    archived: {
        type: DataTypes.BOOLEAN,
        defaultValue: false
    },
    createdAt: {
        type: DataTypes.DATE,
        allowNull: false,
        defaultValue: DataTypes.NOW
    },
    modifiedAt: {
        type: DataTypes.DATE,
        allowNull: true
    }
}, {
    tableName: 'Notes',
    timestamps: false 
});